Move scale and frequency computation to top level
